v2.4.1.2 -> v2.4.1.3
  - Jens: better ordering of requests when unable to merge
  - Neil Brown: make md work as a module again (we cannot autodetect
  in modules, not enough background information)
  - Neil Brown: raid5 SMP locking cleanups
  - Neil Brown: nfsd: handle Irix NFS clients named pipe behavior and
  dentry leak fix
  - maestro3 shutdown fix
  - fix dcache hash calculation that could cause bad hashes under certain
  circumstances (Dean Gaudet)
  - David Miller: networking and sparc updates
  - Jeff Garzik: include file cleanups
  - Andy Grover: ACPI update
  - Coda-fs error return fixes
  - rth: alpha Jensen update

+++ net/khttpd/sysctl.c
@@ -25,7 +25,7 @@ Sysctl interface
 
 #include <linux/kernel.h>
 #include <linux/errno.h>
-#include <linux/malloc.h>
+#include <linux/slab.h>
 #include <linux/net.h>
 #include <linux/sched.h>
 #include <linux/skbuff.h>
@@ -259,7 +259,7 @@ void EndSysctl(void)
 static int proc_dosecurestring(ctl_table *table, int write, struct file *filp,
 		  void *buffer, size_t *lenp)
 {
-	int len;
+	size_t len;
 	char *p, c=0;
 	char String[256];
